VNINDEX SUMM/requirements.txt
flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
playwright==1.40.0
google-cloud-storage==2.13.0
